Initial work on the MVP
Some issues; config takes the entrypoint/binary name as the config filename. So, for a binary called `ds` it will expected a config yaml file at `$HOME/.config/ds/config.yaml`. This is a little problematic for binaries which might have different names.

When deploying the binary name should be consistent so this problem should be avoided.